requirements.txt
anyio==3.3.2
argcomplete==1.12.3
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backpack-for-pytorch==1.3.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
configparser==5.0.2
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.3.2
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
importlib-metadata==4.6.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.4
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.1
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.1
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
requests-unixsocket==0.2.0
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.8.0
sentry-sdk==1.3.1
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
subprocess32==3.5.4
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
tikzplotlib==0.9.15
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.2
traitlets==5.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wandb==0.11.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
zipp==3.5.0
